+++ src/melcloudatw.js
@@ -23,7 +23,6 @@ class MelCloudAtw extends EventEmitter {
 
         //set default values
         this.devicesData = {};
-        this.headers = {};
 
         //lock flags
         this.locks = {
@@ -55,19 +54,21 @@ class MelCloudAtw extends EventEmitter {
         try {
             //read device info from file
             const devicesData = await this.functions.readData(this.devicesFile, true);
-
             if (!Array.isArray(devicesData)) {
                 if (this.logWarn) this.emit('warn', `Device data not found`);
                 return null;
             }
             const deviceData = devicesData.find(device => device.DeviceID === this.deviceId);
-            this.headers = deviceData.Headers;
 
             if (this.accountType === 'melcloudhome') {
                 deviceData.SerialNumber = deviceData.DeviceID || '4.0.0';
                 deviceData.Device.FirmwareAppVersion = deviceData.ConnectedInterfaceIdentifier || '4.0.0';
             }
-            if (this.logDebug) this.emit('debug', `Device Data: ${JSON.stringify(deviceData, null, 2)}`);
+            const safeConfig = {
+                ...deviceData,
+                headers: 'removed',
+            };
+            if (this.logDebug) this.emit('debug', `Device Data: ${JSON.stringify(safeConfig, null, 2)}`);
 
             //device
             const serialNumber = deviceData.SerialNumber;
@@ -152,7 +153,7 @@ class MelCloudAtw extends EventEmitter {
                         method: 'POST',
                         baseURL: ApiUrls.BaseURL,
                         timeout: 10000,
-                        headers: this.headers,
+                        headers: deviceData.Headers,
                         withCredentials: true
                     });
 
@@ -190,11 +191,11 @@ class MelCloudAtw extends EventEmitter {
                         method: 'PUT',
                         baseURL: ApiUrlsHome.BaseURL,
                         timeout: 10000,
-                        headers: this.headers,
+                        headers: deviceData.Headers,
                         withCredentials: true
                     });
 
-                    const settings = {
+                    const settings = effectiveFlags === 'scheduleset' ? { data: { enabled: deviceData.ScheduleEnabled } } : {
                         data: {
                             Power: deviceData.Device.Power,
                             SetTemperatureZone1: deviceData.Device.SetTemperatureZone1,
@@ -213,7 +214,7 @@ class MelCloudAtw extends EventEmitter {
                     };
                     if (this.logDebug) this.emit('debug', `Send Data: ${JSON.stringify(settings.data, null, 2)}`);
 
-                    const path = ApiUrlsHome.SetAtw.replace('deviceid', deviceData.DeviceID);
+                    const path = effectiveFlags === 'scheduleset' ? ApiUrlsHome.SetSchedule.replace('deviceid', deviceData.DeviceID) : ApiUrlsHome.SetAtw.replace('deviceid', deviceData.DeviceID);
                     await axiosInstancePut(path, settings);
                     this.updateData(deviceData);
                     return true;
@@ -221,6 +222,11 @@ class MelCloudAtw extends EventEmitter {
                     return;
             }
         } catch (error) {
+            // Return 500 for schedule hovewer working correct
+            if (error?.response?.status === 500) {
+                return true;
+            }
+            
             throw new Error(`Send data error: ${error.message}`);
         }
     }